Fusion also for the api
