+++ src/api.js
@@ -3,7 +3,7 @@ const deco = require('deco');
 const semver = require('semver');
 const express = require('express');
 const RestError = require('rest-error');
-const Controller = require('../controller');
+const Controller = require('./controller');
 
 // __Module Definition__
 const Api = deco(function(options, protect) {
@@ -58,9 +58,46 @@ const Api = deco(function(options, protect) {
     api.add(controller);
     return controller;
   };
+
+  const controllers = [];
+
+  // __Public Instance Members__
+  // Add a controller to the API.
+  api.add = function(controller) {
+    controllers.push(controller);
+    return api;
+  };
+  // Return a copy of the controllers array, optionally filtered by release.
+  protect.controllers = function(release, fragment) {
+    const all = [].concat(controllers);
+
+    if (!release) return all;
+
+    const satisfies = all.filter(function(controller) {
+      return semver.satisfies(release, controller.versions());
+    });
+
+    if (!fragment) {
+      return satisfies;
+    }
+
+    // Find the matching controller among controllers that match the requested release.
+    return satisfies.filter(function(controller) {
+      return fragment === controller.fragment();
+    });
+  };
+  // Find the correct controller to handle the request.
+  api.use('/:path', function(request, response, next) {
+    const fragment = `/${request.params.path}`;
+    const controllers = protect.controllers(request.baucis.release, fragment);
+    // If not found, bail.
+    if (controllers.length === 0) return next();
+
+    request.baucis.controller = controllers[0];
+    request.baucis.controller(request, response, next);
+  });
 });
 
 Api.factory(express.Router);
-Api.decorators(__dirname, ['controllers']);
 
 module.exports = Api;
