+++ src/controller.js
@@ -1,4 +1,3 @@
-
 const util = require('util');
 const crypto = require('crypto');
 const domain = require('domain');
